+++ 001-core/src/main/java/ru/otus/project/security/dto/UserDto.java
@@ -1,4 +1,4 @@
-package ru.otus.project.dto;
+package ru.otus.project.security.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
