+++ 001-core/src/main/java/ru/otus/project/security/exceptions/EntityNotFoundException.java
@@ -1,0 +1,7 @@
+package ru.otus.project.security.exceptions;
+
+public class EntityNotFoundException extends LibrarySecurityException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
