+++ 001-core/src/main/java/ru/otus/project/security/exceptions/LibrarySecurityException.java
@@ -1,0 +1,7 @@
+package ru.otus.project.security.exceptions;
+
+public class LibrarySecurityException extends RuntimeException {
+    public LibrarySecurityException(String message) {
+        super(message);
+    }
+}
