+++ 001-core/src/main/java/ru/otus/project/security/exceptions/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package ru.otus.project.exceptions;
+package ru.otus.project.security.exceptions;
 
 public class UserNotFoundException extends EntityNotFoundException {
     public UserNotFoundException(String message) {
