+++ 001-core/src/main/java/ru/otus/project/security/services/UserService.java
@@ -1,0 +1,8 @@
+package ru.otus.project.security.services;
+
+import ru.otus.project.security.dto.UserDto;
+
+public interface UserService {
+
+    UserDto findByUsername(String username);
+}
