--- 001-core/src/main/java/ru/otus/project/services/UserService.java
@@ -1,8 +1,0 @@
-package ru.otus.project.services;
-
-import ru.otus.project.dto.UserDto;
-
-public interface UserService {
-
-    UserDto findByUsername(String username);
-}
