+++ 001-core/src/test/java/ru/otus/project/controllers/rest/ControllersSecurityTest.java
@@ -12,7 +12,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import ru.otus.project.config.SecurityConfig;
+import ru.otus.project.security.config.SecurityConfig;
 import ru.otus.project.services.AuthorService;
 import ru.otus.project.services.BookService;
 import ru.otus.project.services.CommentService;
