+++ 001-core/src/test/java/ru/otus/project/security/repositories/JpaUserRepositoryTest.java
@@ -1,11 +1,11 @@
-package ru.otus.project.repositories;
+package ru.otus.project.security.repositories;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import ru.otus.project.models.User;
+import ru.otus.project.security.models.User;
 
 import static org.assertj.core.api.Assertions.assertThat;
